Fix description of direction for mui-slide
The description of the direction parameter for mui-slide states that this would be the side to slide from (`top`, `bottom` etc.). Did not work for me though ... after a short tour through the sources it turns out that it is actually the direction to *slide toward*, i.e., it takes the values `up`, `right`, `down` and `left`.