feat: change mui-series behavior for better Safari support #97
Do not set the animation name until it need to be played. This is the only way to control the animation start on all browsers (including latest Safari versions). See issue for more informations.

**BREAKING CHANGE**
When `.is-animating` is removed from `.mui-series`, the animation is now reset instead of paused. Setting `.is-animating` back will start the animation from its begining.

Closes https://github.com/zurb/motion-ui/issues/97