fix: use polyfill for call() to fix keep support of both Sass 3 and Sass 4
In order to improve modular namespacing, LibSass 4 only accepts first-class functions as argument so functions are called in their own context. In most case, `get-function()` must only be used by the user in its own context.  End developer must be encouraged to use first-class functions.

In this library, we need to use it to keep support of function name strings in Sass > 3.5.

Changes:
* Add `-mui-is-function` and `-mui-safe-call` to test and call function in a safe way for all Sass versions.
* Use `-mui-is-function` instead of a simple `string` type check as functions can now be first-hand functions with a `function` type
* Use `-mui-safe-call` to handle both first-hand functions and function string names the same way. A warning is thrown for function string names with Sass >= 3.5